+++ src/ui/wxWidgets/wxUtilities.cpp
@@ -336,6 +336,24 @@ enum wxUtilities::WindowSystem wxUtilities::WhatWindowSystem()
   return wsType;
 }
 
+bool IsXWaylandEnabled()
+{
+  bool XWayland = false;
+  wxOperatingSystemId osid = wxGetOsVersion();
+
+  if (osid & wxOS_UNIX) { // Includes Linux
+    wxString GDK_BACKEND_VAR = wxEmptyString;
+    if (wxGetEnv(wxT("GDK_BACKEND"), &GDK_BACKEND_VAR)) { // provides 'x11' or 'wayland'
+      if (!GDK_BACKEND_VAR.IsEmpty()) {
+        if (GDK_BACKEND_VAR == wxT("x11")) {
+          XWayland = true;
+        }
+      }
+    }
+  }
+  return XWayland;
+}
+
 bool wxUtilities::IsVirtualKeyboardSupported()
 {
 #ifdef __WINDOWS__
@@ -343,20 +361,23 @@ bool wxUtilities::IsVirtualKeyboardSupported()
 #elif defined __WXOSX__
   return true;
 #else
-  return (wxUtilities::WhatWindowSystem() == wxUtilities::X11);
-#endif
-}
-
-void wxUtilities::DisableIfUnsupported(enum Feature feature, wxWindow* window)
-{
-  if (feature == Autotype && WhatWindowSystem() == Wayland) {
-    window->Disable();
-    window->SetToolTip(_("Not supported by Wayland"));
+  if (wxUtilities::WhatWindowSystem() == wxUtilities::X11) {
+    return true;
   }
+  else if (wxUtilities::WhatWindowSystem() == wxUtilities::Wayland && IsXWaylandEnabled()) {
+    return true;
+  }
+  else {
+    return false;
+  }
+#endif
 }
 
 void wxUtilities::NotifyIfUnsupported(enum Feature feature, wxWindow* window)
 {
+  if (feature == Autotype && WhatWindowSystem() == Wayland) {
+    window->SetToolTip(_("Running on Wayland - make sure the target application is using Xwayland. For some applications, you may need to enable the 'Use alternate AutoType method' option in Options > System."));
+  }
 #if defined(__linux__) || defined(__FreeBSD__) || defined(__OpenBSD__)
   if (feature == SystemTray && !IsTaskBarIconAvailable()) {
     window->SetToolTip(_("Not supported by the current Windowing System (e.g. Wayland), or you may need to install a Desktop Environment extension to enable System Tray support."));